+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Storing hashed paths using `λ` as delimiter instead of `/` to enable some edge cases ([@nivekmai](https://github.com/nivekmai)).
 * Fixing an issue with cursors where a stopped history wouldn't restart correctly ([@nikvm](https://github.com/nikvm)).
 * Fixing monkeys' laziness.
+* Fixing watching paths beneath monkeys.
 
 ## v2.1.1